pyproject.toml
[project]
name = "cp-detector"
version = "0.1.0"
description = "Custom crease pattern detection model for origami"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # Core ML
    "torch>=2.0.0",
    "torchvision>=0.15.0",

    # Backbone models
    "timm>=0.9.0",

    # Data processing
    "albumentations>=1.3.0",
    "opencv-python>=4.8.0",
    "scikit-image>=0.21.0",
    "scipy>=1.11.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "pypdfium2>=4.30.0",

    # Training utilities
    "wandb>=0.15.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "tqdm>=4.65.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
graph = [
    # For Phase 2: Graph Neural Networks
    "torch-geometric>=2.3.0",
]

dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.0.270",
    "mypy>=1.3.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.7.0",
]

[project.scripts]
cp-train = "scripts.train:main"
cp-evaluate = "scripts.evaluate:main"
cp-predict = "scripts.inference:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
